Cleaner: Generate timestamp in UTC timezone
When generating the timestamp for the SquasherClean migration, use `Time.zone` rather than `Time` to ensure that the timestamp is in UTC time. Rails migrations' timestamps are in UTC (for example, when running `rails g migration CreateUsers`) and Squasher migrations should also be in UTC, so that migrations appear in the correct order.